requirements.txt
# Core framework
flask>=3.0.0
werkzeug>=3.0.0

# Authentication
authlib>=1.3.0           # Google OAuth 2.0
flask-login>=0.6.3       # Session management

# Database ORM
flask-sqlalchemy>=3.1.1
flask-migrate>=4.0.5     # DB migrations
sqlalchemy>=2.0.0

# Security
flask-wtf>=1.2.1         # CSRF protection
itsdangerous>=2.1.2      # Secure tokens

# AI / LLM
langchain-groq>=0.1.0
groq>=0.4.0

# File processing
pymupdf>=1.23.0          # PDF reading
python-docx>=1.1.0       # DOCX reading
Pillow>=10.0.0           # Image processing / preview
pytesseract>=0.3.10      # OCR for images

# Export formats
openpyxl>=3.1.0          # Excel export

# Environment & config
python-dotenv>=1.0.0

# HTTP
requests>=2.31.0

# File storage (S3-ready)
boto3>=1.34.0            # AWS S3 (optional, only used when STORAGE_BACKEND=s3)

# Utilities
python-dateutil>=2.8.2
